Part1

+++ 3P2.cpp
@@ -70,7 +70,7 @@ int main(int argc, char *argv[])
 		if (first & 1 == 0) {
 			first += prime;
 		}
-		for (uint64_t i = first; i < size; i += (prime * 2))
+		for (uint64_t i = first; i < size; i += (prime * static_cast<uint64_t>(2)))
 			if (i & 1)//only if i is odd
 				marked[i >> 1] = 1;
 		if (id == 0) {
